refactor: move constants to ParserUtil
