--- src/main/java/seedu/address/logic/parser/ParserConstants.java
@@ -1,13 +1,0 @@
-package seedu.address.logic.parser;
-
-/**
- * Container for constants used in the parser classes.
- */
-public class ParserConstants {
-    public static final String FIELD_NAME = "name";
-    public static final String FIELD_PHONE = "phone";
-    public static final String FIELD_EMAIL = "email";
-    public static final String FIELD_ADDRESS = "address";
-    public static final String COMMA_SEPARATOR = ", ";
-
-}
